oscar branch y eficiencia de código faltante
