+++ huffman_base.py
@@ -1,0 +1,147 @@
+import os
+import sys
+import getopt
+import csv
+from math import log2
+
+# Parametros de entrada y ayuda:
+file_full_path = ""
+file_split_path = []
+def myfunc(argv):
+    global file_full_path, file_split_path
+    arg_output = ""
+    arg_user = ""
+    arg_help = "{0} -i <input>".format(argv[0])
+    
+    try:
+        opts, args = getopt.getopt(argv[1:], "hi:", ["help", "input="])
+    except:
+        print(arg_help)
+        sys.exit(2)
+    
+    for opt, arg in opts:
+        if opt in ("-h", "--help"):
+            print(arg_help)  
+            sys.exit(2)
+        elif opt in ("-i", "--input"):
+            file_full_path = arg
+            file_split_path = os.path.normpath(file_full_path)
+            file_split_path = os.path.split(file_split_path)
+
+
+if __name__ == "__main__":
+    myfunc(sys.argv)
+
+
+file_huffman_comprimido = file_full_path+".huffman"
+ruta_diccionario = file_full_path+".diccionario.csv"
+recovered_path = os.path.join(file_split_path[0], "recovered_"+file_split_path[1])
+#-----------------------------------------------------
+# Algorithmo de compresión de huffman
+#-----------------------------------------------------
+#Apertura y lectura del archivo
+string=[];
+with open(file_full_path, "rb") as f:
+    while (byte := f.read(1)):
+        # Do stuff with byte.
+        int_val = int.from_bytes(byte, "big")
+        string.append(int_val)
+
+# Árbol binario
+class NodeTree(object):
+    def __init__(self, left=None, right=None):
+        self.left = left
+        self.right = right
+    def children(self):
+        return (self.left, self.right)
+    def nodes(self):
+        return (self.left, self.right)
+    def __str__(self):
+        return '%s_%s' % (self.left, self.right)
+
+def insert_in_tree(raiz, ruta, valor):
+    if(len(ruta)==1):
+        if(ruta=='0'):
+            raiz.left = valor;
+        else:
+            raiz.right = valor;
+    else:
+        if(ruta[0]=='0'):
+            #if type(raiz.left) is int:
+            if(raiz.left==None):
+                raiz.left = NodeTree(None,None);
+            ruta = ruta[1:];
+            insert_in_tree(raiz.left,ruta,valor);
+        else:
+            #if type(raiz.right) is int:
+            if(raiz.right==None):
+                raiz.right = NodeTree(None,None);
+            ruta = ruta[1:];
+            insert_in_tree(raiz.right,ruta,valor);
+
+
+# Función principal del algoritmo de Huffman
+def huffman_code_tree(node, left=True, binString=''):
+    if type(node) is int:
+        return {node: binString}
+    (l, r) = node.children()
+    d = dict()
+    d.update(huffman_code_tree(l, True, binString + '0'))
+    d.update(huffman_code_tree(r, False, binString + '1'))
+    return d
+    
+
+# calculo de frecuencias y probabilidades
+prob_unit = 1/len(string)
+freq = {}
+for c in string:
+    if c in freq:
+        freq[c] += prob_unit
+    else:
+        freq[c] = prob_unit
+
+freq = sorted(freq.items(), key=lambda x: x[1], reverse=True)
+
+nodes = freq #Lista de dulpas (símbolo,frecuencia)
+while len(nodes) > 1:
+    (key1, c1) = nodes[-1]
+    (key2, c2) = nodes[-2]
+    nodes = nodes[:-2]
+    node = NodeTree(key1, key2)
+    nodes.append((node, c1 + c2))
+    #print(nodes)
+    nodes = sorted(nodes, key=lambda x: x[1], reverse=True)
+
+huffmanCode = huffman_code_tree(nodes[0][0]) #Diccionario {símbolo:código}
+
+print(' Char | Huffman code ')
+print('----------------------')
+for (char, frequency) in freq:
+    print(' %-4r |%12s' % (char, huffmanCode[char]))
+
+freq=dict(freq) #Diccionario {símbolo:frecuencia}
+
+#Determinar e imprimir la entropía de la fuente:
+information= lambda p:p*log2(1/p)
+entropy=0
+for code in freq:
+    entropy+=information(freq[code])
+print("Entropía de la Fuente: " + str(entropy))
+
+#Determinar e imprimir el largo promedio de los códigos:
+avg_length_new=0
+for code in huffmanCode:
+   avg_length_new+=len(huffmanCode[code])*freq[code]
+print("Largo promedio: " + str(avg_length_new))
+
+#Determinar e imprimir la varianza del nuevo código generado:
+code_variance_new=0
+print("Varianza del código original: " + str(code_variance_new))
+
+#Determinar e imprimir la eficiencia del código original:
+avg_length_old=0
+print("Eficiencia del código original: " + str(avg_length_old))
+
+#Determinar e imprimir la eficiencia del nuevo código generado:
+new_code_efficiency=entropy/avg_length_new
+print("Eficiencia del nuevo código generado: " + str(new_code_efficiency))
